app/requirements.txt
fastapi == 0.115.11
uvicorn == 0.34.0
pydantic == 2.10.6
pydantic-settings == 2.8.1
pydantic[email] == 2.10.6
sqlalchemy == 2.0.39
greenlet == 3.0.3
asyncpg == 0.29.0
psycopg == 3.2.6
psycopg_binary == 3.2.6
passlib == 1.7.4
python-jose[cryptography] == 3.4.0
bcrypt == 4.3.0